requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn==1.6.1
streamlit
xgboost
pandas
pillow
gdown
